+++ src/residential-income-acquisition/index.js
@@ -4,12 +4,14 @@ const contracts = require('./contracts');
 const readiness = require('./readiness');
 const operatingMetrics = require('./operating-metrics');
 const propertyCosts = require('./property-costs');
+const operatingCaseSnapshot = require('./operating-case-snapshot');
 const api = require('./api');
 
-module.exports = Object.assign({}, contracts, readiness, operatingMetrics, propertyCosts, api, {
+module.exports = Object.assign({}, contracts, readiness, operatingMetrics, propertyCosts, operatingCaseSnapshot, api, {
   contracts,
   readiness,
   operatingMetrics,
   propertyCosts,
+  operatingCaseSnapshot,
   api,
 });